+++ test/mix/tasks/ua_inspector/download_test.exs
@@ -54,6 +54,18 @@ defmodule Mix.Tasks.UaInspector.DownloadTest do
     assert String.contains?(console, "Download aborted")
   end
 
+  test "aborted download (quiet)" do
+    Mix.shell(Mix.Shell.IO)
+
+    console =
+      capture_io(fn ->
+        MixTask.run(["--quiet"])
+        IO.write("n")
+      end)
+
+    assert console == "Download databases? [Yn] n"
+  end
+
   test "confirmed download" do
     Mix.shell(Mix.Shell.IO)